Create truthcheker.js
First file I ever post here! Yay? Anyways, it is a function I wrote as a part of getting the FCC certificate in JS and Data Structure, and I am pleased with it. Use at your own discretion, and be aware the copying it directly as your own solution is a direct violation of the FCC Academic Honesty Agreement.